Renamed dropdown and preview folders correctly. Fixed import paths
