+++ src/components/Main/recipeDisplay/RecipeDisplay.tsx
@@ -1,7 +1,8 @@
 import { useContext } from "react"
 import { DataContext } from "../../../context/DataContext"
+import { initRecipeDataState } from "../../../context/initialStates"
 import RecipeItem from "./RecipeItem"
-import { initRecipeDataState } from "../../../context/DataContext"
+
 
 const RecipeDisplay = () => {
     const { recipeData } = useContext(DataContext)
