+++ src/components/Main/selectForm/preview/PreviewListText.tsx
